// src/app/components/enum/constData.enum.ts
export enum EnumConstData{
    profileImageEmpty = 'EmptyImages/User.jpg',
    dbVersionDefault = '0',
    suportTel1 = "۰۲۱-۷۷۷۳۵۲۲۴",
    suportmobile1 = "۰۹۰۳۳۱۴۹۰۸۱",
    suportEmail1 = "info@pulsein.ir",
    suportEmail2 = "support@pulsein.ir",
    addressCo = "تهرانپارس خیابان گلبرگ شرقی نبش خیابان ۱۱۳ مجتمع نگین گلبرگ پلاک ۶۱ واحد ۲",
    textCallToOperator = "برای ارتباط با بخش پشتیبانی پالسین لطفا با شماره‌های زیر تماس بگیرید."
}