// src/app/components/webApp/enum/enumUrls.enum.ts
export enum EnumUrls {
    rulesUrl = "https://pulsein.app/rules",
    address_WebappHome = "Homeeee",
    address_WebappCurrentRequest = "/followRequest",
    address_WebappRequestRegister = "/webApp/requestRegister",
    address_home = "/Home",
    requestRegister_selectJobcat3 = "webApp/requestRegister/selectCategory3",
    requestRegister_extraDescription = "webApp/requestRegister/extraDescription",
    requestRegister_dateTime = "webApp/requestRegister/dateTime",
    requestRegister_location = "webApp/requestRegister/location",
    requestRegister_address = "webApp/requestRegister/address",
    requestRegister_summary = "webApp/requestRegister/summary",
    requestRegister_requestResult = "webApp/requestRegister/requestResult",
    requestRegister_questions = "webApp/requestRegister/questions",
    requestRegister_selectCity = "webApp/requestRegister/selectCity",
    canselrequestPage = "/webApp/cancelRequest",
    requestRegister_loading = "webApp/requestRegister/loading",
}